requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-angular==2.2
django-crispy-forms==1.7.2
django-tastypie==0.14.2
idna==2.8
mysql==0.0.2
mysqlclient==1.4.2.post1
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
